+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes Logs
 
+## v0.2.0
+
+- Added global error handler.
+- Added new trait **TEventListener** for listening and handling event.
+- Added core class **EventBus** to handling the core events such as `error` 
+and `shutdown`.
+
 ## v0.1.1
 
 - Introduce namespace `L\Error` to defines the error codes and segments